+++ src/lib.rs
@@ -7,6 +7,7 @@ pub mod fd;
 pub mod fdio;
 pub mod res;
 pub mod sel;
+pub mod sig;
 pub mod spec;
 pub mod sys;
